requirements.txt
email-validator>=2.2.0
flask-login>=0.6.3
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
pymysql>=1.1.1
paramiko>=3.5.1
jinja2>=3.1.6
cryptography>=44.0.2
sqlalchemy>=2.0.39
werkzeug>=3.1.3
flask-wtf>=1.2.2
dnspython>=2.7.0
python-telegram-bot>=22.0
pytz>=2025.1
requests>=2.32.3
glances>=4.3.1
asyncssh>=2.13.2
aiohttp>=3.9.1
python-dotenv>=1.0.0
